requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
defusedxml==0.8.0rc2
diff-match-patch==20241021
Django==5.1.4
django-allauth==65.3.1
django-celery-beat==2.7.0
django-extensions==3.2.3
django-import-export==4.3.3
django-redis==5.4.0
django-simple-history==3.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
flake8==7.1.1
flower==2.0.1
humanize==4.11.0
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
swagger-spec-validator==3.0.4
tablib==3.7.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
reportlab
